Got rid of repeat functions
functions unnecessarily repeated, got rid of them and just added the code for using the functions. The baseline values are all working. Min value functions are running, but the values they produce are a bit off (same values as before when they were all in separate functions).

Also added new variable names in the for loops but that wasn't the issue.